refactor: move getSystemInfo to verifiers
